Разделил код на 3 модуля
